Add files via upload
There is a render error issue for the first frame. Since the graphic rendering need to sample the compute shader 's output, so the compute command need to run firstly before the graphic rendering.  It is a logical error and can be easy to be reproduced when the GPU device has a lower performance.